+++ h2integrate/control/control_strategies/storage/openloop_storage_control_base.py
