+++ h2integrate/control/control_strategies/storage/simple_openloop_controller.py
@@ -2,7 +2,7 @@ import numpy as np
 from attrs import field, define
 
 from h2integrate.core.utilities import merge_shared_inputs
-from h2integrate.control.control_strategies.storage.openloop_controller_baseclass import (
+from h2integrate.control.control_strategies.storage.openloop_storage_control_base import (
     StorageOpenLoopControlBase,
     StorageOpenLoopControlBaseConfig,
 )
